+++ ouroboros-consensus-cardano/test/cardano-test/Test/ThreadNet/AllegraMary.hs
@@ -240,8 +240,8 @@ prop_simple_allegraMary_convergence TestSetup
                     protocolParamsShelleyBased
                     (SL.ProtVer majorVersion1 0)
                     (SL.ProtVer majorVersion2 0)
-                    ( L.mkTransitionConfig ()
-                    $ L.mkTransitionConfig ()
+                    ( L.mkTransitionConfig L.NoGenesis
+                    $ L.mkTransitionConfig L.NoGenesis
                     $ L.mkShelleyTransitionConfig genesisShelley
                     )
                     hardForkTrigger