+++ sop-extras/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 # Changelog entries
 
+<a id='changelog-0.2.1.0'></a>
+## 0.2.0.0 — 2024-08-26
+
+### Non-Breaking
+
+- Bump to `nothunks` 0.2
+
 <a id='changelog-0.2.0.0'></a>
 ## 0.2.0.0 — 2024-05-13