+++ strict-sop-core/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 # Changelog entries
 
+<a id='changelog-0.1.2.0'></a>
+## 0.1.1.0 — 2024-08-26
+
+### Non-Breaking
+
+- Bump to `nothunks` 0.2
+
 <a id='changelog-0.1.1.0'></a>
 ## 0.1.1.0 — 2024-05-13